Fix Helm chart installing CRDs twice
This was caused by a manual error when creating the Helm chart. In the
future, that error will not happen as we have automation to publish the
chart. In beta.1's case, that automation failed because of a bug that is
now fixed so the chart was published manually.

Fixes #239