Remove unneeded delete optimization
The optimization is faulty because there is no guarantee that the client
caches are up to date. It is possible for delete of a child resource to
happen when the parent resource isn't in the cache. This is especially
probably immediately after pod restart or if the parent has the
reconcile-action: skip annotation and has already been deleted.

Most of the rest of the PR is re-recording tests and samples.