Refactor cross resource reference config overrides
The original cross resource reference config override handler classified
properties as references (or not). This works OK for properties of type
*string or string but doesn't work for collections like []string or
map[string]Something, because in those cases it's not the property that
is the reference it's the type (or a part of the type).

The Swagger "format": "arm-id" importer already used the astmodel.ARMID
type. Now the config based overrides also use the same type and work for
collections.