Make the operator pod manage CRDs (#2769)
Adds a new command line parameter to the operator pod, allowing configuration of what CRDs to install. Initially only supports "*" (or empty).

"*" means install all the CRDs.
unspecified means don't install any CRDs

We thought about using a CRD to configure this, but there were a number of downsides:

- It required extra work to kubectl apply the operator because the CRD creation races with the instance of CRD. This significantly complicated initial install (1 step to 3)
- It required extra work to helm install. The extra work could be hidden inside helm for the most part, but it did require helm --wait - if you didn't use --wait installation would error with a somewhat obscure error.

I considered moving to a ConfigMap, but given our desired goal of a clear failure mode, it actually made the most sense to move to a pod cmdline parameter. Changing the parameters in the deployment will automatically trigger a cycle of the underlying pods. Using a ConfigMap but trying to put the status onto the pod was awkward as you have to launch the ctrl.Manager first in order to start the reconciler, at which point the pod will start responding to /healthz as healthy.

Another advantage of the cmdline approach is that it's promising the least. If we have a CRD or ConfigMap which we're watching and reconciling they might be surprised that any update triggers a pod restart. Making it a cmdline parameter makes it very clear that the operator pod will have to restart to apply the new changes. If we want to evolve to promise more in the future based on user requests, we can but we're not committing to it up front.

Also adds a /readyz probe to ensure that the pod is ready. In this case ready includes CRD reconciliation, so the pod will not be marked as ready until it has listed current CRDs and possibly added new ones.

Helm chart testing:

- Default configuration - Passes
- installCRDs = false - Confirmed we only install + use the crd-reader role. Pod starts successfully but watches 0 resources (because 0 resources installed). For users to get this to work, they can manually install whatever CRDs they want BEFORE chart and operator will dynamically discover them.
- Multitenant standard - Confirmed it works, updated Taskfile test to use Helm rather than raw YAML as this actually didn't work before.
- Multitenant CRDs off - Confirmed we use crd-reader role
- Upgrade from previous chart - passes

Future possible improvements:
- Support more than just *.