Split ARMClientCache into CredentialProvider and Cache (#2947)
* Move arm_client_cache to internal/identity

* Split ARMClientCache into CredentialProvider and Cache

* A separate CredentialProvider encapsulates logic for looking up the
  specific credential.
* ARMClientCache now only concerns itself with caching logic for
  clients. Actual credential lookup is deferred to the
  CredentialProvider.

Prior to this change, the assumption was that every credential was
always used to produce a GenericARMClient. This enables scenarios
where the user for the credential is something other than an ARM client.
One example of such a usage is to extract an AAD token to authenticate
with the MySQL Dataplane.