Add support for MySQL FlexibleServer AAD Users (#2972)
Includes support for creating AAD users.

In order to manage AAD users, the identity the operator is running as
(either the global identity or the identity specified in the users
credential-from annotation) must be the AAD Administrator of the MySQL
server. This requires that ASO is using a managed identity of some kind.
We recommend workload identity.