Update_index.md
This PR is to improvise the Developer Setup (with VS Code) steps. 

In order to complete the clone, the tags are required in local. in some cases, the cloning the repo does not automatically clone the tags, which results in generating an error message. 

If we miss the added step and run `task` command, it shows error:
Error Message: task: Command "/workspaces/azure-service-operator/scripts/v2/build_version.py v2" failed: exit status1. 

By performing the added step, one will be saved from fall for unnecessary setup issues.