Fix networking resources deletion of child resources during adoption (#3136)
Networking resources have a structure where a PUT on the parent also can delete child resources (or add them). ASO attempts to allow users to manage these networking child resources independent of the owner. In order to do that, when building the PUT payload for the parent resource we used the ownerReferences.uid field to look up the child resources owned by the parent resource we were about to send and include them all in the PUT payload of the owner so that they are not deleted.

The issue is that with adoption, parent + child already exist in Azure, but ownerReferences.uid isn't set yet in Kubernetes, which means the PUT of the parent doesn't include any reference to its children and so PUT parent ends up deleting the children.

This issue impacts the following resources:

VirtualNetwork
LoadBalancer
RouteTable
NetworkSecurityGroup
We investigated using .spec.owner instead of ownerReferences, which works, but inconsistently as there is still a race in Kubernetes client caching when the parent is created during adoption. Sometimes it realizes that it has children and sometimes it doesn't.

Instead, we rely on Azure itself, issuing a GET parent while building the PUT payload. We use the exact contents of the GET response as the PUT payload for the child resource field (subnets for VNET, rules for LoadBalancer, etc). This ensures that we don't ever delete any child resource when PUT-ing the parent, even if we're adopting that parent for the first time.

Note that technically GET + PUT is subject to a dual-writer race, which we could use ETag and If-Match headers to prevent. We didn't implement that here because:

Some of the networking APIs support this paradigm but it's not universally available.
It would require moderate changes to the reconciler to pipe the etag through.
ASO expects to own the resource entirely, so if there is another parallel writer their changes will inevitably lose to ASO as it will keep updating the Azure resource to its desired state.
Given the above it didn't make sense to bother with optimistic concurrency.