Fix v3 recording infrastructure
Fix bug where interaction.Request.RequestURI was used when
interaction.Request.URL should have been used. RequestURI seems to
always be empty.

Fix bug where body hashes were computed before subscriptionID and
other Azure IDs were redacted, resulting in a hash that doesn't match
the body that was eventually recorded (that body had these values
redacted).
This issue only showed up for requests that contained ARM IDs in the
body which is why we didn't notice it before.
The issue here was that there are two contexts that use the
vcr/redact.go helpers, one during replay and one
during recording. For replay, there's no need to redact subID because we
set it to 0000's, but during recording (such as when we compute the body
hash header) we need to redact the subids first.