WIP: Attempt to make ASOv1 CI work without an SP
This is pushed as-is in case we need to complete this work at some
point.

Note that the publishing parts of the CI pipeline would probably
work if the TODOs in azure-pipelines.yml were completed. The
ServiceConnections to the prerelease and release ACRs have already been
created.

The harder part is testing either locally or in CI, as both cases seem
to have assumptions baked in that rely on ServicePrincipals we would
need to correct (or disable?). If you're under time pressure to publish
an ASOv1 release I would recommend just disabling the KeyVault test
stuff, fixing what we need to fix, running the tests locally, and then
reviving the publishing workflow only (comment everything else out in
the CI job).