Change how we read configuration
Moving from a model where we have to explicitly set each secret in the
pod env to a model where we just mount the whole secret into the pod.
This has the added benefit of allowing us to autodetect secret changes
and restart the pod AND means that users can totally omit the secret now
as opposed to having to create the secret with empty values.