Fix bug where owner.armId could be changed (#4193)
It was never intended to allow changes to owner.armId, this protection
was missed when adding owner.armId support.

Even though this is technically a breaking change in the strictest
sense, it's unlikely to break users as updating ARM ID, while not
explicitly rejected, didn't actually work upstream in Azure for most
cases.

Fixes #4181