Add envtest panic handler (#4404)
We disable controller-runtime logging for our envtest instances because
we create multiple of them and it can be a lot of extra logs we don't
need, but we rely on controller-runtime logs for panic-detection.

We want panic logging with controller-runtime logging disabled for
envtest, but we don't want duplicate panic logs in production, so we
introduce a PanicHandler that is set only in envtest contexts.