Fix bug preventing resource reconciliation due to "Secret not cached" (#4966)
This was caused by PR #4857 setting ReaderFailOnMissingInformer. It
happens only if the CRDs installed via crdPattern do not include a resource
which requires a Secret informer.

This issue was missed in our testing because we generally run with a
full (or close to full) set of CRDs installed, but the problem only
appears with a minimal subset installed.

Fixes #4926.