Fix select annotation changed predicate (#4967)
It was mistakenly classifying creates as containing an annotation
change. This was previously harmless because we always reconciled the
resource (say, ResourceGroup) if the obj was created anyway, but now
that we also use this predicate for watching namespaces, we do not want
events on namespace create, only on namespace updates that contain an
annotation change.