Fix taskfile ordering bug with zz_generated.go files (#4974)
We run the full header-check script at v2/ directory for many tasks
running in parallel. This can cause issues where the controller task
deletes the zz_generated.go files after the check_headers script has
discovered them via file-listing, which causes the script to error
because the file was deleted out from under it.

Fix to only run check_headers on the directory in quesiton. This also
has the added benefit of reducing duplication of checks slightly (we
still have some because controller checks all of v2 which covers
generator and asoctl anyway).